+++ src/rest/rest.container.ts
@@ -5,7 +5,7 @@ import { PinoLogger, Logger } from '../shared/libs/logger/index.js';
 import { DatabaseClient, MongoDatabaseClient } from '../shared/libs/database-client/index.js';
 import { Component } from '../shared/types/component.enum.js';
 import { RestApplication } from './rest.application.js';
-import { AppExceptionFilter, ExceptionFilter } from '../shared/libs/rest/index.js';
+import { AppExceptionFilter, ExceptionFilter, HttpErrorExceptionFilter, ValidationExceptionFilter } from '../shared/libs/rest/index.js';
 
 export function createRestApplicationContainer() {
   const container = new Container();
@@ -14,5 +14,7 @@ export function createRestApplicationContainer() {
   container.bind<Config<RestSchemaType>>(Component.Config).to(RestConfig).inSingletonScope();
   container.bind<DatabaseClient>(Component.DatabaseClient).to(MongoDatabaseClient).inSingletonScope();
   container.bind<ExceptionFilter>(Component.ExceptionFilter).to(AppExceptionFilter).inSingletonScope();
+  container.bind<ExceptionFilter>(Component.HttpExceptionFilter).to(HttpErrorExceptionFilter).inSingletonScope();
+  container.bind<ExceptionFilter>(Component.ValidationExceptionFilter).to(ValidationExceptionFilter).inSingletonScope();
   return container;
 }
