+++ src/shared/libs/rest/exception-filter/http-error.exception-filter.ts
@@ -1,0 +1,28 @@
+import { NextFunction, Request, Response } from 'express';
+import { inject, injectable } from 'inversify';
+import { ExceptionFilter } from './exception-filter.interface.js';
+import { Component } from '../../../types/component.enum.js';
+import { Logger } from '../../logger/index.js';
+import { StatusCodes } from 'http-status-codes';
+import { createErrorObject } from '../../../helpers/index.js';
+import { HttpError } from '../index.js';
+import { ApplicationError } from '../../../../const.js';
+
+@injectable()
+export class HttpErrorExceptionFilter implements ExceptionFilter {
+  constructor(
+    @inject(Component.Logger) private readonly logger: Logger
+  ) {
+    this.logger.info('Register HttpErrorExceptionFilter');
+  }
+
+  public catch(error: unknown, req: Request, res: Response, next: NextFunction): void {
+    if (error instanceof HttpError) {
+      return next(error);
+    }
+
+    this.logger.error(`[HttpErrorException]: ${req.path} # ${error.message}`, error);
+
+    res.status(StatusCodes.BAD_REQUEST).json(createErrorObject(ApplicationError.CommonError, error.message));
+  }
+}
