+++ src/shared/libs/rest/index.ts
@@ -1,8 +1,10 @@
 export * from './controller/base-controller.abstract.js';
 export * from './controller/controller.interface.js';
 export * from './exception-filter/exception-filter.interface.js';
+export * from './exception-filter/app.exception-filter.js';
+export * from './exception-filter/http-error.exception-filter.js';
+export * from './exception-filter/validation.exception-filter.js';
 export * from './route/route.interface.js';
-export * from './exception-filter/app-exception-filter.js';
 export * from './errors/http-error.js';
 export * from './middleware/middleware.interface.js';
 export * from './middleware/validate-dto.middleware.js';
